requirements.txt
plotly == 4.5.2
tensorboard == 2.4.0
test-tube == 0.7.5
Pillow == 8.4.0
PyYAML == 5.3.1
joblib == 0.15.1
matplotlib == 3.4.2
mlflow == 1.17
numpy == 1.20.1
opencv_python == 4.4.0.44
pandas == 1.1.5
pytorch_lightning == 1.1.8
scikit_image == 0.16.2
sh == 1.14.2
sklearn == 0.0
torch == 1.8.1
torchmetrics == 0.3.2
torchvision == 0.9.1
tqdm == 4.63
attrdict == 2.0.1
pywise == 0.3.2
shapely == 1.8.1
cityscapesscripts~=2.2.0
